Added nyan_cat class. Added the spritesheet for nyan cat. DrawNyanCat function that draws the animated nyan cat going accross the screen.

+++ src/location.cpp
@@ -6,7 +6,7 @@
 
 namespace space_invader {
 
-Location::Location(int row, int col) : row_(row), col_(col) {}
+Location::Location(double row, double col) : row_(row), col_(col) {}
 
 bool Location::operator==(const Location& rhs) const {
   return row_ == rhs.row_ && col_ == rhs.col_;
@@ -52,7 +52,7 @@ Location& Location::operator+=(const Location& rhs) {
 }
 
 
-int mod(int a, int b) {
+double mod(int a, int b) {
   int c = a % b;
   return c + (c < 0 ? b : 0);
 }
@@ -61,9 +61,9 @@ Location Location::operator%(const Location& rhs) const {
   return {mod(row_, rhs.row_), mod(col_, rhs.col_)};
 }
 
-int Location::Row() const { return row_; }
+double Location::Row() const { return row_; }
 
-int Location::Col() const { return col_; }
+double Location::Col() const { return col_; }
 
 
 }
